Remove duplicated fixtures.
